Moved util/Threading.h to thread/Threading.h and renamed the namespace
