+++ openvdb/openvdb/CMakeLists.txt
@@ -434,8 +434,11 @@ set(OPENVDB_LIBRARY_UTIL_INCLUDE_FILES
   util/NodeMasks.h
   util/NullInterrupter.h
   util/PagedArray.h
-  util/Threading.h
   util/Util.h
+)
+
+set(OPENVDB_LIBRARY_THREAD_INCLUDE_FILES
+  thread/Threading.h
 )
 
 # @todo CMake >= 3.12, use an object library to consolidate shared/static
@@ -593,3 +596,4 @@ install(FILES ${OPENVDB_LIBRARY_POINTS_INCLUDE_FILES} DESTINATION ${OPENVDB_INST
 install(FILES ${OPENVDB_LIBRARY_TOOLS_INCLUDE_FILES} DESTINATION ${OPENVDB_INSTALL_INCLUDEDIR}/tools)
 install(FILES ${OPENVDB_LIBRARY_TREE_INCLUDE_FILES} DESTINATION ${OPENVDB_INSTALL_INCLUDEDIR}/tree)
 install(FILES ${OPENVDB_LIBRARY_UTIL_INCLUDE_FILES} DESTINATION ${OPENVDB_INSTALL_INCLUDEDIR}/util)
+install(FILES ${OPENVDB_LIBRARY_THREAD_INCLUDE_FILES} DESTINATION ${OPENVDB_INSTALL_INCLUDEDIR}/thread)
