+++ openvdb/openvdb/tools/ParticlesToLevelSet.h
@@ -69,7 +69,7 @@
 #include "openvdb/tree/LeafManager.h"
 #include "openvdb/util/logging.h"
 #include "openvdb/util/NullInterrupter.h"
-#include "openvdb/util/Threading.h"
+#include "openvdb/thread/Threading.h"
 
 #include "Composite.h" // for csgUnion()
 #include "PointPartitioner.h"
@@ -742,7 +742,7 @@ private:
         for (Coord c = lo; c.x() <= hi.x(); ++c.x()) {
             //only check interrupter every 32'th scan in x
             if (!(count++ & ((1<<5)-1)) && util::wasInterrupted(mParent.mInterrupter)) {
-                util::cancelGroupExecution();
+                thread::cancelGroupExecution();
                 return false;
             }
             const Real x2 = math::Pow2(c.x() - P[0]);
@@ -808,7 +808,7 @@ private:
         // Densely fill the remaining regions.
         for (const auto& bbox: padding) {
             if (util::wasInterrupted(mParent.mInterrupter)) {
-                util::cancelGroupExecution();
+                thread::cancelGroupExecution();
                 return false;
             }
             const Coord &bmin = bbox.min(), &bmax = bbox.max();
